feat: adding file paging to collections. Partial implementation

+++ src/main/java/org/techhouse/data/DbEntry.java
@@ -15,7 +15,7 @@ public class DbEntry {
     private String databaseName;
     private String collectionName;
     private JsonObject data;
-    private long page;
+    private int page;
 
     public static DbEntry fromJsonObject(String databaseName, String collectionName, JsonObject jsonObject) {
         final var entry = new DbEntry();
